feat: add option to toggle quoting the git message
The fix for issue #215 adds double quotes to the git message prior to
calling git commit -m. However, there are situations where the message
ends up like this:

    "docs: Update README.md"

instead of:

    docs: Update README.md

There's also situations where if you don't double quote the message,
execa() does this and causes an error:

    git commit -m docs: Update README.md

Instead of this:

    git commit -m "docs: Update README.md"

For now, just add the option to add double quotes or not.